refactor: remove arithmetics and autograd_functions from base_modules
