+++ tests/base_modules/arithmetics/test_binary_arithmetics.py
@@ -1,44 +1,42 @@
 import pytest
 import torch
 
-from elasticai.creator.base_modules.arithmetics.binary_arithmetics import (
-    BinaryArithmetics,
-)
+from elasticai.creator.nn.binary._math_operations import MathOperations
 
-arithmetic = BinaryArithmetics()
+operations = MathOperations()
 
 
 def test_quantization_returns_1_for_0_8():
     x = torch.tensor([0.8])
-    x = arithmetic.quantize(x)
+    x = operations.quantize(x)
     x = x.tolist()
     assert x == [1.0]
 
 
 def test_quantization_returns_1_for_2_3():
     x = torch.tensor([2.3])
-    x = arithmetic.quantize(x)
+    x = operations.quantize(x)
     x = x.tolist()
     assert x == [1.0]
 
 
 def test_quantization_returns_minus_1_for_minus_0_7():
     x = torch.tensor([-0.7])
-    x = arithmetic.quantize(x)
+    x = operations.quantize(x)
     x = x.tolist()
     assert x == [-1.0]
 
 
 def test_quantization_returns_minus_1_for_minus_3():
     x = torch.tensor([-3])
-    x = arithmetic.quantize(x)
+    x = operations.quantize(x)
     x = x.tolist()
     assert x == [-1.0]
 
 
 def test_quantization_produces_gradient_1_for_minus_0_3():
     x = torch.tensor([-0.3], requires_grad=True)
-    y = arithmetic.quantize(x)
+    y = operations.quantize(x)
     y.backward()
     g = x.grad
     assert g.tolist() == [1.0]
@@ -46,7 +44,7 @@ def test_quantization_produces_gradient_1_for_minus_0_3():
 
 def test_quantization_produces_gradient_0_for_3():
     x = torch.tensor([3.0], requires_grad=True)
-    y = arithmetic.quantize(x)
+    y = operations.quantize(x)
     y.backward()
     g = x.grad
     assert g.tolist() == [0.0]
@@ -63,7 +61,7 @@ def test_quantization_produces_gradient_0_for_3():
 def test_addition(a, b, result):
     x = torch.tensor([float(a)])
     y = torch.tensor([float(b)])
-    assert arithmetic.add(x, y) == torch.tensor([float(result)])
+    assert operations.add(x, y) == torch.tensor([float(result)])
 
 
 @pytest.fixture(params=((-1, -1, 1), (1, 1, -1), (-1, -1, -1), (1, 1, 1)))
@@ -76,20 +74,11 @@ def operants_as_tensors(operants) -> torch.Tensor:
     return torch.tensor([float(x) for x in operants])
 
 
-def test_summation(operants, operants_as_tensors):
-    number_of_ones = len(tuple((filter(lambda x: x > 0, operants))))
-    expected = -1.0
-    if number_of_ones > len(operants) / 2:
-        expected = 1.0
-    result = arithmetic.sum(operants_as_tensors).tolist()
-    assert result == expected
-
-
 def test_matmul():
     a = [[-1.0, -1.0], [1.0, -1]]
     b = [[1.0, 1.0], [-1.0, 1.0]]
     expected = [[1.0, -1.0], [1.0, 1.0]]
-    actual = arithmetic.matmul(torch.tensor(a), torch.tensor(b)).tolist()
+    actual = operations.matmul(torch.tensor(a), torch.tensor(b)).tolist()
     assert actual == expected
 
 
@@ -97,5 +86,5 @@ def test_mul():
     a = [-1.0, -1.0, 1.0, 1.0]
     b = [1.0, -1.0, -1.0, 1.0]
     expected = [-1.0, 1.0, -1.0, 1.0]
-    actual = arithmetic.mul(torch.tensor(a), torch.tensor(b)).tolist()
+    actual = operations.mul(torch.tensor(a), torch.tensor(b)).tolist()
     assert actual == expected
