+++ tests/base_modules/arithmetics/test_binary_arithmetics_with_layers.py
@@ -1,9 +1,7 @@
 import torch
 
-from elasticai.creator.base_modules.arithmetics.binary_arithmetics import (
-    BinaryArithmetics,
-)
 from elasticai.creator.base_modules.linear import Linear
+from elasticai.creator.nn.binary._math_operations import MathOperations
 
 
 class TestLinearLayer:
@@ -12,7 +10,7 @@ class TestLinearLayer:
         weight = [[0.1, -0.8], [-1.2, 4.0]]
         expected = [-1.0, 1.0]
         linear = Linear(
-            operations=BinaryArithmetics(),
+            operations=MathOperations(),
             in_features=len(x),
             out_features=len(weight),
             bias=False,
