+++ tests/base_modules/arithmetics/test_float_arithmetics.py
@@ -1,31 +1,11 @@
 import torch
 
-from elasticai.creator.base_modules.arithmetics.float_arithmetics import (
-    FloatArithmetics,
-)
+from elasticai.creator.nn.float._math_operations import MathOperations
 from tests.tensor_test_case import assertTensorEqual
 
 
-def test_clamp() -> None:
-    arithmetics = FloatArithmetics(mantissa_bits=3, exponent_bits=1)
-    a = torch.tensor([-3.0, -2.0, -1.0, 0.0, 1.0, 2.0, 3.0])
-    assertTensorEqual(
-        expected=[-1.875, -1.875, -1.0, 0.0, 1.0, 1.875, 1.875],
-        actual=arithmetics.clamp(a),
-    )
-
-
-def test_round() -> None:
-    arithmetics = FloatArithmetics(mantissa_bits=3, exponent_bits=1)
-    a = torch.tensor([-1.69, -0.2, 0.2, 1.69])
-    assertTensorEqual(
-        expected=[-1.75, -0.203125, 0.203125, 1.75],
-        actual=arithmetics.round(a),
-    )
-
-
 def test_quantize() -> None:
-    arithmetics = FloatArithmetics(mantissa_bits=3, exponent_bits=1)
+    arithmetics = MathOperations(mantissa_bits=3, exponent_bits=1)
     a = torch.tensor([-3.0, -2.0, -1.69, -0.2, 0.2, 1.69, 2.0, 3.0])
     assertTensorEqual(
         expected=[-1.875, -1.875, -1.75, -0.203125, 0.203125, 1.75, 1.875, 1.875],
@@ -34,7 +14,7 @@ def test_quantize() -> None:
 
 
 def test_add() -> None:
-    arithmetics = FloatArithmetics(mantissa_bits=3, exponent_bits=1)
+    arithmetics = MathOperations(mantissa_bits=3, exponent_bits=1)
     a = torch.tensor([-0.6875, 0.1250, 0.6875])
     b = torch.tensor([-0.3125, 0.2031, 0.3125])
     assertTensorEqual(
@@ -43,17 +23,8 @@ def test_add() -> None:
     )
 
 
-def test_sum() -> None:
-    arithmetics = FloatArithmetics(mantissa_bits=3, exponent_bits=1)
-    a = torch.tensor([-1.0, 0.3125, 1.0])
-    assertTensorEqual(
-        expected=0.3125,
-        actual=arithmetics.sum(a),
-    )
-
-
 def test_mul() -> None:
-    arithmetics = FloatArithmetics(mantissa_bits=3, exponent_bits=1)
+    arithmetics = MathOperations(mantissa_bits=3, exponent_bits=1)
     a = torch.tensor([-0.6875, 0.1250, 0.6875])
     b = torch.tensor([-0.3125, 0.2031, 0.3125])
     assertTensorEqual(
@@ -63,27 +34,10 @@ def test_mul() -> None:
 
 
 def test_matmul() -> None:
-    arithmetics = FloatArithmetics(mantissa_bits=3, exponent_bits=1)
+    arithmetics = MathOperations(mantissa_bits=3, exponent_bits=1)
     a = torch.tensor([[-2.0, -1.75, -1.5], [-0.25, 0.0, 0.25], [1.25, 1.5, 1.75]])
     b = torch.tensor([[-0.25], [0.5], [0.25]])
     assertTensorEqual(
         expected=[[-0.75], [0.125], [0.875]],
         actual=arithmetics.matmul(a, b),
-    )
-
-
-def test_conv1d() -> None:
-    arithmetics = FloatArithmetics(mantissa_bits=3, exponent_bits=1)
-    a = torch.tensor([[-1.75, -1.5, -1, -0.25, 1, 2.5, 3.75]])
-    assertTensorEqual(
-        expected=[[-1.875, -1.5, -0.25, 1.75, 1.875, 1.875]],
-        actual=arithmetics.conv1d(
-            inputs=a,
-            weights=torch.ones(1, 1, 2),
-            bias=torch.ones(1),
-            stride=1,
-            padding="valid",
-            dilation=1,
-            groups=1,
-        ),
     )
